Header Image

+++ TunnelVision/app/src/main/java/com/way/tunnelvision/entity/service/HeaderImageService.java
@@ -12,7 +12,7 @@ import com.way.tunnelvision.R;
 import com.way.tunnelvision.base.MainApp;
 import com.way.tunnelvision.entity.impl.HeaderImageModelImpl;
 import com.way.tunnelvision.entity.model.HeaderImageModel;
-import com.way.tunnelvision.ui.activity.MainActivity;
+import com.way.tunnelvision.ui.activity.BackgroundActivity;
 import com.way.tunnelvision.util.LogUtil;
 
 import java.util.List;
@@ -81,24 +81,29 @@ public class HeaderImageService extends Service {
                             if (list != null && list.size() > 0) {
                                 LogUtil.d(TAG, "onStartCommand debug, HeaderImageModels COUNT = " + list.size());
                                 HeaderImageDaoHelper headerImageDaoHelper = HeaderImageDaoHelper.getInstance();
-                                headerImageDaoHelper.deleteAll();
-                                for (HeaderImageModel item : list) {
-                                    headerImageDaoHelper.addData(item);
-                                }
+                                HeaderImageModel firstItem = list.get(0);
+                                if(headerImageDaoHelper.hasEntity(firstItem)){
+                                    LogUtil.d(TAG, "onStartCommand debug, No Need to Update");
+                                } else {
+                                    headerImageDaoHelper.deleteAll();
+                                    for (HeaderImageModel item : list) {
+                                        headerImageDaoHelper.addData(item);
+                                    }
 
-                                NotificationManager mn = (NotificationManager) HeaderImageService.this.getSystemService(NOTIFICATION_SERVICE);
-                                Notification.Builder builder = new Notification.Builder(HeaderImageService.this);
-                                Intent notificationIntent = new Intent(HeaderImageService.this, MainActivity.class);//点击跳转位置
-                                PendingIntent contentIntent = PendingIntent.getActivity(HeaderImageService.this, 0, notificationIntent, 0);
-                                builder.setContentIntent(contentIntent);
-                                builder.setSmallIcon(R.mipmap.ic_launcher);
-                                builder.setTicker(intent.getStringExtra("tickerText")); //测试通知栏标题
-                                builder.setContentText(intent.getStringExtra("contentText")); //下拉通知啦内容
-                                builder.setContentTitle(intent.getStringExtra("contentTitle"));//下拉通知栏标题
-                                builder.setAutoCancel(true);
-                                builder.setDefaults(Notification.DEFAULT_ALL);
-                                Notification notification = builder.build();
-                                mn.notify((int) System.currentTimeMillis(), notification);
+                                    NotificationManager mn = (NotificationManager) HeaderImageService.this.getSystemService(NOTIFICATION_SERVICE);
+                                    Notification.Builder builder = new Notification.Builder(HeaderImageService.this);
+                                    Intent notificationIntent = new Intent(HeaderImageService.this, BackgroundActivity.class);//点击跳转位置
+                                    PendingIntent contentIntent = PendingIntent.getActivity(HeaderImageService.this, 0, notificationIntent, 0);
+                                    builder.setContentIntent(contentIntent);
+                                    builder.setSmallIcon(R.mipmap.ic_launcher);
+                                    builder.setTicker(intent.getStringExtra("tickerText")); //测试通知栏标题
+                                    builder.setContentText(intent.getStringExtra("contentText")); //下拉通知啦内容
+                                    builder.setContentTitle(intent.getStringExtra("contentTitle"));//下拉通知栏标题
+                                    builder.setAutoCancel(true);
+                                    builder.setDefaults(Notification.DEFAULT_ALL);
+                                    Notification notification = builder.build();
+                                    mn.notify((int) System.currentTimeMillis(), notification);
+                                }
                             }
                         }
